# packages/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/boards/{boardId}/fider-board": {
      "post": {
        "operationId": "FiderBoardController_createOrUpdate",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "The ID of the board to set",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FiderBoardCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FiderBoard"
                }
              }
            }
          }
        },
        "summary": "Create or update a Fider board",
        "tags": ["Fider Boards"]
      },
      "get": {
        "operationId": "FiderBoardController_getByBoardId",
        "parameters": [
          {
            "name": "boardId",
            "required": true,
            "in": "path",
            "description": "The ID of the board to get",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Get a Fider board by board ID",
        "tags": ["Fider Boards"]
      }
    }
  },
  "info": {
    "title": "BoardBot API",
    "description": "BoardBot API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FiderBoardCreateDto": {
        "type": "object",
        "properties": {
          "baseUrl": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          }
        },
        "required": ["baseUrl", "apiKey"]
      },
      "FiderBoard": {
        "type": "object",
        "properties": {}
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}